feat(react-router): nested persistLayouts and graph-boundary split
Cache each layout tier independently when persistLayouts is enabled, skip
refreshPersistedLayout on initial bootstrap sync, and extract client graph
boundary reachability and AST transforms into dedicated modules.
